major refactor, allow 3 types of webview
