+++ extensions/void/src/webviews/common/posthog.tsx
