+++ extensions/void/src/webviews/ctrlk/index.tsx
@@ -1,0 +1,7 @@
+import React from "react"
+import { mount } from "../common/mount"
+import { CtrlK } from "./CtrlK"
+
+// this is the entry point that mounts ctrlk
+mount(<CtrlK />)
+
