+++ extensions/void/src/webviews/sidebar/SidebarSettings.tsx
@@ -1,5 +1,5 @@
 import React, { useState } from "react";
-import { configFields, useVoidConfig, VoidConfigField } from "./contextForConfig";
+import { configFields, useVoidConfig, VoidConfigField } from "../common/contextForConfig";
 
 
 const SettingOfFieldAndParam = ({ field, param }: { field: VoidConfigField, param: string }) => {
