+++ extensions/void/src/webviews/sidebar/index.tsx
@@ -1,0 +1,7 @@
+import React from "react"
+import Sidebar from "./Sidebar"
+import { mount } from "../common/mount"
+
+// this is the entry point that mounts the sidebar
+mount(<Sidebar />)
+
